tests: tfm_regression_test: Adapt to Split build
The non-secure test application has to be build separately in the
tf-m-tests repository as an external project.

The regression tests need to provide test configurations to both
TF-M an NS app.
Duplicate configuration done in the spe/CMakeLists.txt to configure
TF-M image for the regression tests.
While coping some files from TF-M to the non-secure interface to make
the tests run

Signed-off-by: Markus Swarowsky <markus.swarowsky@nordicsemi.no>