nrf_security: Use MBEDTLS_PSA_CRYPTO_CONFIG_FILE
So far the MBEDTLS_USER_CONFIG_FILE was used to pass the PSA
configuration to mbedTLS but with mbed TLS 3.5.2 it has its own
MBEDTLS_PSA_CRYPTO_CONFIG_FILE so changing the build system to use that
for passing the PSA_WANT configs which is now required for legacy and
PSA crypto builds.

The MBEDTLS_PSA_CRYPTO_USER_CONFIG_FILE is now used for the PSA_NEED
definitions.

Signed-off-by: Markus Swarowsky <markus.swarowsky@nordicsemi.no>